hoàn thiện phần player

+++ src/main/java/Entity/Player.java
@@ -38,7 +38,7 @@ public class Player extends Entity{
     }
 
     //cơ chế trừ máu
-    public void TakeDame(int damage) {
+    public void TakeDame(double damage) {
         this.hp -= damage;
         if(this.hp <=0 ) {
             this.hp =0;
